+++ samplesheets/forms.py
@@ -419,3 +419,22 @@ class IrodsRequestAcceptForm(forms.Form):
         label='I accept the iRODS delete request',
         required=True,
     )
+
+
+class SampleSheetVersionCompareForm(forms.Form):
+    """Form for selecting two sheet versions for comparison."""
+
+    def __init__(self, choices, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields['source'] = forms.CharField(
+            label='Source',
+            widget=forms.Select(
+                choices=choices, attrs={'class': 'form-control'}
+            ),
+        )
+        self.fields['target'] = forms.CharField(
+            label='Target',
+            widget=forms.Select(
+                choices=choices, attrs={'class': 'form-control'}
+            ),
+        )